pyproject.toml
[tool.poetry]
name = "hyperion"
version = "0.1.0.post10"
description = "A headless ETL / ELT / data pipeline and integration SDK for Python."
authors = ["Tomas Votava <info@tomasvotava.eu>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
loguru = "^0.7.2"
fastavro = "^1.9.7"
boto3 = "^1.35.59"
env-proxy = "^1.1.0"
python-dotenv = "^1.0.1"
python-snappy = "^0.7.3"
googlemaps = "^4.10.0"
cachetools = "^5.5.0"
python-dateutil = "^2.9.0.post0"
pydantic = "^2.10.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.3"
mypy = "^1.13.0"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
boto3-stubs = "^1.35.59"
types-cachetools = "^5.5.0.20240820"
types-python-dateutil = "^2.9.0.20241003"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "S101",   # allow asserts in tests
    "RUF018", # allow asserts with walrus in tests
]


[tool.ruff.lint]
select = [
    #"D",
    "E",
    "F",
    "B",
    "I",
    "N",
    "UP",
    "S",
    "A",
    "DTZ",
    "PT",
    "SIM",
    "PTH",
    "PD",
    "RUF",
    "T20",
]

[tool.pytest.ini_options]
addopts = [
    "-vv",
    "--cov=hyperion",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-report=lcov",
    "tests/",
]

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = ["snappy.*", "googlemaps.*"]
ignore_missing_imports = true
